grammar: handle composition text when committing grammar suggestions.
When composition is enabled, DeleteSurroundingText may not delete the
text as we expected, so in the case we use SetCompositionRange to
select the error text then CommitText will overwrite it with grammar
suggestion.

Bug: 1225538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2999902
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#897998}

+++ chrome/browser/chromeos/input_method/grammar_manager.cc
@@ -225,24 +225,32 @@ void GrammarManager::AcceptSuggestion() {
     LOG(ERROR) << "Failed to commit grammar suggestion.";
   }
 
-  // NOTE: GetSurroundingTextInfo() could return a stale cache that no
-  // longer reflects reality, due to async-ness between IMF and
-  // TextInputClient.
-  // TODO(crbug/1194424): Work around the issue or fix
-  // GetSurroundingTextInfo().
-  const ui::SurroundingTextInfo surrounding_text =
-      input_context->GetSurroundingTextInfo();
+  if (input_context->HasCompositionText()) {
+    input_context->SetComposingRange(current_fragment_.range.start(),
+                                     current_fragment_.range.end(), {});
+    input_context->CommitText(
+        base::UTF8ToUTF16(current_fragment_.suggestion),
+        ui::TextInputClient::InsertTextCursorBehavior::kMoveCursorAfterText);
+  } else {
+    // NOTE: GetSurroundingTextInfo() could return a stale cache that no
+    // longer reflects reality, due to async-ness between IMF and
+    // TextInputClient.
+    // TODO(crbug/1194424): Work around the issue or fix
+    // GetSurroundingTextInfo().
+    const ui::SurroundingTextInfo surrounding_text =
+        input_context->GetSurroundingTextInfo();
 
-  // Delete the incorrect grammar fragment.
-  input_context->DeleteSurroundingText(
-      -static_cast<int>(surrounding_text.selection_range.start() -
-                        current_fragment_.range.start()),
-      current_fragment_.range.length() -
-          surrounding_text.selection_range.length());
-  // Insert the suggestion and put cursor after it.
-  input_context->CommitText(
-      base::UTF8ToUTF16(current_fragment_.suggestion),
-      ui::TextInputClient::InsertTextCursorBehavior::kMoveCursorAfterText);
+    // Delete the incorrect grammar fragment.
+    input_context->DeleteSurroundingText(
+        -static_cast<int>(surrounding_text.selection_range.start() -
+                          current_fragment_.range.start()),
+        current_fragment_.range.length() -
+            surrounding_text.selection_range.length());
+    // Insert the suggestion and put cursor after it.
+    input_context->CommitText(
+        base::UTF8ToUTF16(current_fragment_.suggestion),
+        ui::TextInputClient::InsertTextCursorBehavior::kMoveCursorAfterText);
+  }
 
   RecordGrammarAction(GrammarActions::kAccepted);
 }
